refactor: MenuGroup 을 JPA 엔티티로 변경, Repository 적용
