+++ src/main/java/kitchenpos/domain/menugroup/MenuGroupRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.domain.menugroup;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+
+}
